remove links from copied response
