+++ ui/src/pages/RagChatTab/ChatOutput/ChatMessages/ChatMessageBody.tsx
@@ -36,25 +36,17 @@
  * DATA.
  */
 
-import {
-  ChatMessageType,
-  CrewEventResponse,
-  SourceNode,
-} from "src/api/chatApi.ts";
+import { ChatMessageType, CrewEventResponse } from "src/api/chatApi.ts";
 import UserQuestion from "pages/RagChatTab/ChatOutput/ChatMessages/UserQuestion.tsx";
 import { Divider, Flex, Typography } from "antd";
 import Images from "src/components/images/Images.ts";
 import { cdlBlue500, cdlGray200 } from "src/cuix/variables.ts";
-import Markdown from "react-markdown";
-import Remark from "remark-gfm";
 import { Evaluations } from "pages/RagChatTab/ChatOutput/ChatMessages/Evaluations.tsx";
 import RatingFeedbackWrapper from "pages/RagChatTab/ChatOutput/ChatMessages/RatingFeedbackWrapper.tsx";
 import CopyButton from "pages/RagChatTab/ChatOutput/ChatMessages/CopyButton.tsx";
 import StreamedEvents from "pages/RagChatTab/ChatOutput/ChatMessages/StreamedEvents.tsx";
-import rehypeRaw from "rehype-raw";
-import { SourceCard } from "pages/RagChatTab/ChatOutput/Sources/SourceCard.tsx";
-import { ComponentProps, ReactElement } from "react";
 import SourceNodes from "pages/RagChatTab/ChatOutput/Sources/SourceNodes.tsx";
+import { MarkdownResponse } from "pages/RagChatTab/ChatOutput/ChatMessages/MarkdownResponse.tsx";
 
 export const ChatMessageBody = ({
   data,
@@ -102,45 +94,7 @@ export const ChatMessageBody = ({
             <Flex vertical gap={8} style={{ width: "100%" }}>
               <StreamedEvents streamedEvents={streamedEvents} />
               <Typography.Text style={{ fontSize: 16, marginTop: 8 }}>
-                <Markdown
-                  // skipHtml={true}
-                  remarkPlugins={[Remark]}
-                  rehypePlugins={[rehypeRaw]}
-                  className="styled-markdown"
-                  children={data.rag_message.assistant.trimStart()}
-                  components={{
-                    a: (
-                      props: ComponentProps<"a">,
-                    ): ReactElement<SourceNode> | undefined => {
-                      const { href, className, children, ...other } = props;
-                      if (className === "rag_citation") {
-                        if (data.source_nodes.length === 0) {
-                          return undefined;
-                        }
-                        const { source_nodes } = data;
-                        const sourceNodeIndex = source_nodes.findIndex(
-                          (source_node) => source_node.node_id === href,
-                        );
-                        if (sourceNodeIndex >= 0) {
-                          return (
-                            <SourceCard
-                              source={source_nodes[sourceNodeIndex]}
-                              index={sourceNodeIndex + 1}
-                            />
-                          );
-                        }
-                        if (!href?.startsWith("http")) {
-                          return undefined;
-                        }
-                      }
-                      return (
-                        <a href={href} className={className} {...other}>
-                          {children}
-                        </a>
-                      );
-                    },
-                  }}
-                />
+                <MarkdownResponse data={data} />
               </Typography.Text>
               <SourceNodes data={data} />
               <Flex gap={16} align="center">
